added demo.ipynb and updated readme
