+++ README.md
@@ -11,15 +11,15 @@ This repository containes code and data for the paper 'Exploring replay' current
 The repository is structured as follows:
 - `paper/code` contains all models and maze configurations
 - - `paper/code/bandit` contains code for the bandit simulations
-  - `paper/code/maze` contains code for the Tolman maze simulations
+  - `paper/code/maze` contains code for the Tolman maze simulations. There is also a jupyter notebook `demo.ipynb` which shows how to run the code as well as the expected output.
   - `paper/code/mazes` contains maze configurations
-- `paper/figures_code` contains scripts which can be run to generate the data and any individual figure in the manuscript. Each individual figure has its own folder. Note that the figures were additionally assembled using Inkscape.
+- `paper/figures_code` contains scripts which can be run to generate the data and any individual figure in the manuscript. Each individual figure has its own folder. Note that the figures were additionally assembled using Inkscape
 
 # Software requirements
 ## OS requirements
 The code was developed and tested on Linux Manjaro (5.10.186-1)
 ## Python dependencies 
-All the code was written in `Python 3.9.5` with the following external packages: 
+All the code was written in `Python 3.9.7` with the following external packages: 
 ```
 numpy 
 matplotlib
