+++ paper/figures_code/supp/supp3/supp_3_plot.py
@@ -4,33 +4,45 @@ import sys, os, pickle
 sys.path.append(os.path.abspath(os.path.join(sys.path[0], '../../../code/maze')))
 from utils import plot_maze
 
-load_path = os.path.abspath(os.path.join(sys.path[0], '../../../figures/supp/supp3/data/1/0'))
-save_path = os.path.abspath(os.path.join(sys.path[0], '../../../figures/supp/supp3/'))
+load_path = os.path.abspath(os.path.join(sys.path[0], '../../../figures/supp/supp2/data/'))
+save_path = os.path.abspath(os.path.join(sys.path[0], '../../../figures/supp/supp3'))
 
 def main():
 
-    with open(os.path.join(load_path, 'ag.pkl'), 'rb') as f:
+    with open(os.path.join(load_path, '0', '0', 'ag.pkl'), 'rb') as f:
         agent = pickle.load(f)
 
-    fig = plt.figure(figsize=(14, 4), constrained_layout=True, dpi=100)
+    priors = [[2, 2], [6, 2], [10, 2], [14, 2], [18, 2], [22, 2]]
+    betas  = [1, 2, 4, 'greedy']
+    sas    = [[14, 0], [20, 0], [19, 3], [18, 3], [24, 0], [30, 0], [31, 2], [32, 2]]
+    probas = np.ones((len(priors), len(betas)))
 
-    ax1 = fig.add_subplot(131)
-    plot_maze(ax1, np.load(os.path.join(load_path, 'q_mb.npy')), agent, colorbar=True, colormap='Purples', move=[38])
-    ax1.set_title(r'Initial behavioural policy', fontsize=16)
-    # ax1.set_ylabel(r'Initial $Q^{MF}$', fontsize=14)
-    ax1.text(-0.1, 1.1, 'A', transform=ax1.transAxes, fontsize=16, fontweight='bold', va='top', ha='right')
+    for idxb, beta in enumerate(betas):
+        agent.beta = beta
+        for idxp, prior in enumerate(priors):
 
-    ax2 = fig.add_subplot(132)
-    q_explore_replay_diff = np.load(os.path.join(load_path, 'q_explore_replay_diff.npy'))
-    q_explore_replay_diff[q_explore_replay_diff == 0.] = np.nan
-    plot_maze(ax2, q_explore_replay_diff, agent, colorbar=True, colormap='Purples')
-    ax2.set_title(r'Exploratory replay', fontsize=16)
-    ax2.text(-0.1, 1.1, 'B', transform=ax2.transAxes, fontsize=16, fontweight='bold', va='top', ha='right')
+            this_path = os.path.join(load_path, str(idxb), str(idxp))
 
-    ax3 = fig.add_subplot(133)
-    plot_maze(ax3, np.load(os.path.join(load_path, 'q_explore_replay.npy')), agent, colorbar=True, colormap='Purples', move=[38])
-    ax3.set_title(r'Updated exploratory policy', fontsize=16)
-    ax3.text(-0.1, 1.1, 'C', transform=ax3.transAxes, fontsize=16, fontweight='bold', va='top', ha='right')
+            Q = np.load(os.path.join(this_path, 'q_explore_replay.npy'))
+
+            for sa in sas:
+
+                s, a = sa[0], sa[1]
+
+                probas[idxp, idxb] *= agent._policy(Q[s, :])[a]
+
+    fig = plt.figure(figsize=(4, 3), constrained_layout=True, dpi=100)
+
+    colours = ['blue', 'orange', 'green', 'purple', 'red']
+
+    for idxb, beta in enumerate(betas):
+        plt.plot(range(len(priors)), probas[:, idxb], c=colours[idxb], label=r'$\beta=$%s'%beta)
+
+    plt.legend(prop={'size':8})
+    plt.ylabel('Exploration probability', fontsize=14)
+    # plt.xticks(range(len(priors)), ['Beta(' + str(i).strip('[').strip(']') + ')' for i in priors], rotation=45)
+    plt.xticks(range(len(priors)), [np.round(i[0]/(i[0]+i[1]), 2) for i in priors], rotation=45)
+    plt.xlabel(r'$\mathbb{E}_b[p(open)]$', fontsize=14)
 
     plt.savefig(os.path.join(save_path, 'supp_3.png'))
     plt.savefig(os.path.join(save_path, 'supp_3.svg'), transparent=True)
